+++ spec/datadog/appsec/metrics/exporter_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Datadog::AppSec::Metrics::Exporter do
     let(:span) { instance_double(Datadog::Tracing::SpanOperation) }
     let(:empty_store) do
       Datadog::AppSec::Metrics::Collector::Store.new(
-        evals: 0, timeouts: 0, duration_ns: 0, duration_ext_ns: 0
+        evals: 0, timeouts: 0, duration_ns: 0, duration_ext_ns: 0, downstream_requests: 0
       )
     end
 
@@ -29,7 +29,7 @@ RSpec.describe Datadog::AppSec::Metrics::Exporter do
     context 'when waf metrics present' do
       let(:waf_store) do
         Datadog::AppSec::Metrics::Collector::Store.new(
-          evals: 10, timeouts: 5, duration_ns: 1000, duration_ext_ns: 2000
+          evals: 10, timeouts: 5, duration_ns: 1000, duration_ext_ns: 2000, downstream_requests: 0
         )
       end
 
@@ -45,7 +45,7 @@ RSpec.describe Datadog::AppSec::Metrics::Exporter do
     context 'when RASP metrics present' do
       let(:rasp_store) do
         Datadog::AppSec::Metrics::Collector::Store.new(
-          evals: 10, timeouts: 5, duration_ns: 1000, duration_ext_ns: 2000
+          evals: 10, timeouts: 5, duration_ns: 1000, duration_ext_ns: 2000, downstream_requests: 0
         )
       end
 
@@ -62,7 +62,7 @@ RSpec.describe Datadog::AppSec::Metrics::Exporter do
     context 'when RASP metrics present with no timeouts' do
       let(:rasp_store) do
         Datadog::AppSec::Metrics::Collector::Store.new(
-          evals: 3, timeouts: 0, duration_ns: 1000, duration_ext_ns: 2000
+          evals: 3, timeouts: 0, duration_ns: 1000, duration_ext_ns: 2000, downstream_requests: 0
         )
       end
 
@@ -70,6 +70,23 @@ RSpec.describe Datadog::AppSec::Metrics::Exporter do
         expect(span).to receive(:set_tag).with('_dd.appsec.rasp.rule.eval', 3)
         expect(span).to receive(:set_tag).with('_dd.appsec.rasp.duration', 1.0)
         expect(span).to receive(:set_tag).with('_dd.appsec.rasp.duration_ext', 2.0)
+
+        described_class.export_rasp_metrics(rasp_store, span)
+      end
+    end
+
+    context 'when RASP downstream requests present' do
+      let(:rasp_store) do
+        Datadog::AppSec::Metrics::Collector::Store.new(
+          evals: 3, timeouts: 0, duration_ns: 1000, duration_ext_ns: 2000, downstream_requests: 2
+        )
+      end
+
+      it 'sets downstream requests metric on the span' do
+        expect(span).to receive(:set_tag).with('_dd.appsec.rasp.rule.eval', 3)
+        expect(span).to receive(:set_tag).with('_dd.appsec.rasp.duration', 1.0)
+        expect(span).to receive(:set_tag).with('_dd.appsec.rasp.duration_ext', 2.0)
+        expect(span).to receive(:set_tag).with('_dd.appsec.downstream_request', 2)
 
         described_class.export_rasp_metrics(rasp_store, span)
       end