+++ spec/datadog/profiling/component_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe Datadog::Profiling::Component do
     if Datadog::Profiling.supported?
       allow(Datadog::Profiling::Tasks::Setup).to receive(:new).and_return(profiler_setup_task)
       allow(Datadog::Profiling::Ext::DirMonkeyPatches).to receive(:apply!).and_return(true)
+      settings.profiling.advanced.shutdown_on_exec_enabled = false
     end
   end
 
@@ -144,9 +145,10 @@ RSpec.describe Datadog::Profiling::Component do
           build_profiler_component
         end
 
-        context "when gc_enabled is true" do
+        context "when gc_enabled is true", ruby: '>= 2.7' do
           before do
             settings.profiling.advanced.gc_enabled = true
+
             stub_const("RUBY_VERSION", testing_version)
           end
 
@@ -210,6 +212,11 @@ RSpec.describe Datadog::Profiling::Component do
           before do
             settings.profiling.allocation_enabled = true
             settings.profiling.advanced.gc_enabled = false # Disable this to avoid any additional warnings coming from it
+
+            # Since RUBY_VERSION is used to test if the ExecMonkeyPatch should be applied, mocking it below on an old
+            # Ruby would cause it to wrongly be triggered, so we avoid running these specs there.
+            skip "Behavior does not apply to current Ruby version" if RUBY_VERSION < "2.7" && testing_version >= "3"
+
             stub_const("RUBY_VERSION", testing_version)
           end
 
@@ -308,7 +315,7 @@ RSpec.describe Datadog::Profiling::Component do
           end
         end
 
-        context "when heap profiling is enabled" do
+        context "when heap profiling is enabled", ruby: '>= 2.7' do
           # Universally supported ruby version for allocation profiling by default
           let(:testing_version) { "3.3.0" }
 
@@ -646,6 +653,43 @@ RSpec.describe Datadog::Profiling::Component do
 
           it "is not applied" do
             expect(Datadog::Profiling::Ext::DirMonkeyPatches).to_not receive(:apply!)
+
+            build_profiler_component
+          end
+        end
+      end
+
+      describe "exec workaround" do
+        let(:exec_monkey_patch_name) { "Datadog::Profiling::Ext::ExecMonkeyPatch" }
+
+        context "when can_apply_exec_monkey_patch? is false" do
+          before do
+            allow(described_class).to receive(:can_apply_exec_monkey_patch?).and_return(false)
+          end
+
+          it "does not apply the exec monkey patch" do
+            # Validate there's no previous leaked state
+            expect(Object.ancestors.map(&:to_s)).to_not include(exec_monkey_patch_name)
+
+            build_profiler_component
+
+            expect(Object.ancestors.map(&:to_s)).to_not include(exec_monkey_patch_name)
+          end
+        end
+
+        context "when can_apply_exec_monkey_patch? is true", ruby: '>= 2.7' do
+          let(:exec_monkey_patch) { class_double(exec_monkey_patch_name, apply!: true) }
+
+          before do
+            allow(described_class).to receive(:can_apply_exec_monkey_patch?).and_return(true)
+
+            require "datadog/profiling/ext/exec_monkey_patch" # Make sure it's loaded, so we can mock it cleanly
+
+            stub_const(exec_monkey_patch_name, exec_monkey_patch)
+          end
+
+          it "applies the exec monkey patch" do
+            expect(exec_monkey_patch).to receive(:apply!)
 
             build_profiler_component
           end
@@ -1049,6 +1093,46 @@ RSpec.describe Datadog::Profiling::Component do
       end
 
       include_examples "no_signals_workaround_enabled :auto behavior"
+    end
+  end
+
+  describe ".can_apply_exec_monkey_patch?" do
+    subject(:can_apply_exec_monkey_patch?) { described_class.send(:can_apply_exec_monkey_patch?, settings) }
+
+    context "on Ruby < 2.7" do
+      before { stub_const("RUBY_VERSION", "2.6.9") }
+
+      it "returns false and does not require the monkey patch" do
+        expect(described_class).to_not receive(:require)
+
+        expect(can_apply_exec_monkey_patch?).to be false
+      end
+    end
+
+    context "on Ruby >= 2.7" do
+      before do
+        stub_const("RUBY_VERSION", "2.7.0")
+      end
+
+      context "when exec workaround is disabled" do
+        before { settings.profiling.advanced.shutdown_on_exec_enabled = false }
+
+        it "returns false but still requires the monkey patch" do
+          expect(described_class).to receive(:require).with("datadog/profiling/ext/exec_monkey_patch")
+
+          expect(can_apply_exec_monkey_patch?).to be false
+        end
+      end
+
+      context "when exec workaround is enabled" do
+        before { settings.profiling.advanced.shutdown_on_exec_enabled = true }
+
+        it "returns true and requires the monkey patch" do
+          expect(described_class).to receive(:require).with("datadog/profiling/ext/exec_monkey_patch")
+
+          expect(can_apply_exec_monkey_patch?).to be true
+        end
+      end
     end
   end
 end